requirements.txt
django
djangorestframework
google-generativeai
requests
python-dotenv